add choice for using specialized triples for mul/conv/convbw

+++ image_analysis/pond/tensor.py
@@ -219,7 +219,7 @@ class NativeTensor:
         out = out.transpose(3, 0, 1, 2)
         return out, X_col
 
-    def conv2d_bw(x, d_y, filter_shape, cached_col, precomputed=None, use_specialized_triple=True):
+    def conv2d_bw(x, d_y, filter_shape, cached_col):
         if isinstance(d_y, NativeTensor) or isinstance(d_y, PublicEncodedTensor):
             assert cached_col is not None
             h_filter, w_filter, d_filter, n_filter = filter_shape
@@ -469,7 +469,7 @@ class PublicEncodedTensor:
         return out, X_col
 
 
-    def conv2d_bw(x, d_y, filter_shape, cached_col, precomputed=None, use_specialized_triple=True):
+    def conv2d_bw(x, d_y, filter_shape, cached_col):
         if isinstance(d_y, NativeTensor) or isinstance(d_y, PublicEncodedTensor):
             assert cached_col is not None
             h_filter, w_filter, d_filter, n_filter = filter_shape
@@ -687,17 +687,20 @@ class PrivateFieldTensor:
             return PrivateFieldTensor.from_shares(shares0, shares1)
 
 
-def generate_mul_triple(shape):
-    a = np.array([random.randrange(Q) for _ in range(np.prod(shape))]).astype(DTYPE).reshape(shape)
-    b = np.array([random.randrange(Q) for _ in range(np.prod(shape))]).astype(DTYPE).reshape(shape)
+def generate_mul_triple(shape1, shape2):
+    a = np.array([random.randrange(Q) for _ in range(np.prod(shape1))]).astype(DTYPE).reshape(shape1)
+    b = np.array([random.randrange(Q) for _ in range(np.prod(shape2))]).astype(DTYPE).reshape(shape2)
     ab = (a * b) % Q
     return PrivateFieldTensor.from_elements(a), \
            PrivateFieldTensor.from_elements(b), \
            PrivateFieldTensor.from_elements(ab)
 
-def generate_dot_triple(m, n, o):
-    a = np.array([random.randrange(Q) for _ in range(m * n)]).astype(DTYPE).reshape((m, n))
-    b = np.array([random.randrange(Q) for _ in range(n * o)]).astype(DTYPE).reshape((n, o))
+def generate_dot_triple(m, n, o, a=None, b=None):
+    if a is None: a = np.array([random.randrange(Q) for _ in range(m * n)]).astype(DTYPE).reshape((m, n))
+    else: a = a.reveal().elements
+    if b is None: b = np.array([random.randrange(Q) for _ in range(n * o)]).astype(DTYPE).reshape((n, o))
+    else: b = b.reveal().elements
+
     ab = np.dot(a, b)
     return PrivateFieldTensor.from_elements(a), \
            PrivateFieldTensor.from_elements(b), \
@@ -752,6 +755,10 @@ class PrivateEncodedTensor:
         assert shares0.shape == shares1.shape
         self.shares0 = shares0
         self.shares1 = shares1
+        self.mask = None
+        self.masked_transformed = None
+        self.masked = None
+        self.masked_transformed = None
 
     def from_values(values):
         return PrivateEncodedTensor(values)
@@ -847,29 +854,39 @@ class PrivateEncodedTensor:
     def __sub__(x, y):
         return x.sub(y)
 
-    def mul(x, y, precomputed=None):
+    def mul(x, y, precomputed=None, use_specialized_triple=True):
         y = wrap_if_needed(y)
         if isinstance(y, PublicEncodedTensor):
             shares0 = (x.shares0 * y.elements) % Q
             shares1 = (x.shares1 * y.elements) % Q
             return PrivateEncodedTensor.from_shares(shares0, shares1).truncate()
         if isinstance(y, PrivateEncodedTensor):
-            x_broadcasted0, y_broadcasted0 = np.broadcast_arrays(x.shares0, y.shares0)
-            x_broadcasted1, y_broadcasted1 = np.broadcast_arrays(x.shares1, y.shares1)
-            x_broadcasted = PrivateEncodedTensor.from_shares(x_broadcasted0, x_broadcasted1)
-            y_broadcasted = PrivateEncodedTensor.from_shares(y_broadcasted0, y_broadcasted1)
+            if use_specialized_triple:
+                if precomputed is None: precomputed = generate_mul_triple(x.shape, y.shape)
+                a, b, ab = precomputed
+                alpha = (x - a).reveal()
+                beta = (y - b).reveal()
+                z = alpha.mul(beta) + \
+                    alpha.mul(b) + \
+                    a.mul(beta) + \
+                    ab
+            else:
+                x_broadcasted0, y_broadcasted0 = np.broadcast_arrays(x.shares0, y.shares0)
+                x_broadcasted1, y_broadcasted1 = np.broadcast_arrays(x.shares1, y.shares1)
+                x_broadcasted = PrivateEncodedTensor.from_shares(x_broadcasted0, x_broadcasted1)
+                y_broadcasted = PrivateEncodedTensor.from_shares(y_broadcasted0, y_broadcasted1)
 
-            if precomputed is None: precomputed = generate_mul_triple(x_broadcasted.shape)
-            a, b, ab = precomputed
-            assert x_broadcasted.shape == y_broadcasted.shape
-            assert x_broadcasted.shape == a.shape
-            assert y_broadcasted.shape == b.shape
-            alpha = (x_broadcasted - a).reveal()
-            beta = (y_broadcasted - b).reveal()
-            z = alpha.mul(beta) + \
-                alpha.mul(b) + \
-                a.mul(beta) + \
-                ab
+                if precomputed is None: precomputed = generate_mul_triple(x_broadcasted.shape, y_broadcasted.shape)
+                a, b, ab = precomputed
+                assert x_broadcasted.shape == y_broadcasted.shape
+                assert x_broadcasted.shape == a.shape
+                assert y_broadcasted.shape == b.shape
+                alpha = (x_broadcasted - a).reveal()
+                beta = (y_broadcasted - b).reveal()
+                z = alpha.mul(beta) + \
+                    alpha.mul(b) + \
+                    a.mul(beta) + \
+                    ab
             return PrivateEncodedTensor.from_shares(z.shares0, z.shares1).truncate()
         if isinstance(y, PrivateFieldTensor):
             shares0 = (x.shares0 * y.shares0) % Q
@@ -884,7 +901,7 @@ class PrivateEncodedTensor:
     def __mul__(x, y):
         return x.mul(y)
 
-    def dot(x, y, precomputed=None, save_mask=True):
+    def dot(x, y, precomputed=None, save_mask=False):
         y = wrap_if_needed(y)
         if isinstance(y, PublicEncodedTensor):
             assert x.shape[-1] == y.shape[0]
@@ -896,16 +913,36 @@ class PrivateEncodedTensor:
             n = x.shape[1]
             o = y.shape[1]
             assert n == y.shape[0]
-            if precomputed is None: precomputed = generate_dot_triple(m, n, o)
+            a, b, alpha, beta = None, None, None, None
+            if x.mask is not None:
+                a = x.mask
+                alpha = x.masked
+            if y.mask is not None:
+                b = y.mask
+                beta = y.masked
+
+            if precomputed is None: precomputed = generate_dot_triple(m, n, o, a, b)
             a, b, ab = precomputed
-            alpha = (x - a).reveal() # (PrivateEncodedTensor - PrivateFieldTensor).reveal() = PublicFieldTensor
-            beta = (y - b).reveal()  # (PrivateEncodedTensor - PrivateFieldTensor).reveal() = PublicFieldTensor
+
+            if alpha is None:
+                alpha = (x - a).reveal() # (PrivateEncodedTensor - PrivateFieldTensor).reveal() = PublicFieldTensor
+            if beta is None:
+                beta = (y - b).reveal()  # (PrivateEncodedTensor - PrivateFieldTensor).reveal() = PublicFieldTensor
+
             z = alpha.dot(beta) + alpha.dot(b) + a.dot(beta) + ab
             # PublicFieldTensor.dot(PublicFieldTensor) = PublicFieldTensor
             # PublicFieldTensor.dot(PrivateFieldTensor) = PrivateFieldTensor
             # PrivateFieldTensor.dot(PublicFieldTensor) = PrivateFieldTensor
             # PrivateFieldTensor = PrivateFieldTensor
             # PublicFieldTensor + PrivateFieldTensor + PrivateFieldTensor + PrivateFieldTensor = PrivateFieldTensor
+
+            # cache masks
+            if save_mask:
+                x.mask = a
+                x.masked = alpha
+                if isinstance(y, PrivateEncodedTensor):
+                    y.mask = b
+                    y.masked = beta
             return PrivateEncodedTensor.from_shares(z.shares0, z.shares1).truncate()
         raise TypeError("%s does not support %s" % (type(x), type(y)))
 
@@ -915,23 +952,6 @@ class PrivateEncodedTensor:
         if isinstance(y, PublicEncodedTensor): return x.mul(y.inv())
         raise TypeError("%s does not support %s" % (type(x), type(y)))
 
-    # def matmul(x, y, precomputed=None):
-    #     y = wrap_if_needed(y)
-    #     if isinstance(y, PublicEncodedTensor):
-    #         shares0 = np.matmul(x.shares0, y.elements) % Q
-    #         shares1 = np.matmul(x.shares1, y.elements) % Q
-    #         return PrivateEncodedTensor.from_shares(shares0, shares1).truncate()
-    #     if isinstance(y, PrivateEncodedTensor):
-    #         if precomputed is None: precomputed = generate_matmul_triple(x.shape, y.shape)
-    #         a, b, ab = precomputed
-    #         alpha = (x - a).reveal()
-    #         beta = (y - b).reveal()
-    #         z = np.matmul(alpha, beta) + \
-    #             np.matmul(alpha, b) + \
-    #             np.matmul(a, beta) + \
-    #             ab
-    #         return PrivateEncodedTensor.from_shares(z.shares0, z.shares1).truncate()
-    #     raise TypeError("%s does not support %s" % (type(x), type(y)))
 
     def __truediv__(x, y):
         return x.div(y)
@@ -1002,30 +1022,29 @@ class PrivateEncodedTensor:
             return PrivateEncodedTensor.from_shares(shares0, shares1)
 
 
-    def conv2d(x, filters, strides, padding, use_specialized_triple=True, precomputed=None):
-        h_filter, w_filter, d_filters, n_filters = filters.shape
+    def conv2d(x, y, strides, padding, use_specialized_triple=True, precomputed=None, save_mask=True):
+        h_filter, w_filter, d_y, n_filters = y.shape
         n_x, d_x, h_x, w_x = x.shape
         h_out = int((h_x - h_filter + 2 * padding) / strides + 1)
         w_out = int((w_x - w_filter + 2 * padding) / strides + 1)
 
-        if isinstance(filters, PublicEncodedTensor):
+        if isinstance(y, PublicEncodedTensor):
             X_col = x.im2col(h_filter, w_filter, padding, strides)
-            W_col = filters.transpose(3, 2, 0, 1).reshape(n_filters, -1)
-            out = W_col.dot(X_col).reshape(n_filters, h_out, w_out, n_x).transpose(3, 0, 1, 2)
+            y_col = y.transpose(3, 2, 0, 1).reshape(n_filters, -1)
+            out = y_col.dot(X_col).reshape(n_filters, h_out, w_out, n_x).transpose(3, 0, 1, 2)
             return out, X_col
 
-        if isinstance(filters, PrivateEncodedTensor):
+        if isinstance(y, PrivateEncodedTensor):
             if use_specialized_triple:
-                start = time.time()
                 if precomputed is None:
-                    precomputed = generate_conv_triple(x.shape, filters.shape, strides, padding)
+                    precomputed = generate_conv_triple(x.shape, y.shape, strides, padding)
 
                 # this part of code creates the masks and masked values. a and b are PrivateFieldTensors.
                 # alpha and beta are PublicFieldTensors. The final result is a PrivateFieldTensor
 
                 a, b, a_conv_b, a_col = precomputed          # PrivateFieldTensors
                 alpha = (x - a).reveal()
-                beta = (filters - b).reveal()
+                beta = (y - b).reveal()
 
                 alpha_col = alpha.im2col(h_filter, w_filter, padding, strides)
                 beta_col = beta.transpose(3, 2, 0, 1).reshape(n_filters, -1)
@@ -1037,20 +1056,25 @@ class PrivateEncodedTensor:
 
                 z = (alpha_conv_beta + alpha_conv_b + a_conv_beta + a_conv_b).reshape(n_filters, h_out,
                                                                                       w_out, n_x).transpose(3, 0, 1, 2)
-                return PrivateEncodedTensor.from_shares(z.shares0, z.shares1).truncate(), (a, a_col, alpha_col)
+                # cache
+                if save_mask:
+                    x.mask, x.mask_transformed, x.masked_transformed = a, a_col, alpha_col
+                    y.mask, y.mask_transformed, y.masked_transformed = b, b_col, beta_col
+
+                return PrivateEncodedTensor.from_shares(z.shares0, z.shares1).truncate(), None
 
             else:
                 X_col = x.im2col(h_filter, w_filter, padding, strides)
-                W_col = filters.transpose(3, 2, 0, 1).reshape(n_filters, -1)
+                W_col = y.transpose(3, 2, 0, 1).reshape(n_filters, -1)
                 out = W_col.dot(X_col).reshape(n_filters, h_out, w_out, n_x).transpose(3, 0, 1, 2)
                 return out, X_col
 
-        raise TypeError("%s does not support %s" % (type(x), type(filters)))
+        raise TypeError("%s does not support %s" % (type(x), type(y)))
 
 
 
-    def conv2d_bw(x, d_y, filter_shape, cache, use_specialized_triple=True, reuse_mask=True, precomputed=None):
-        #for Native and PublicEncoded Tensors we cache X_col, for PrivateEncodedTensors we cache alpha_col
+    def conv2d_bw(x, d_y, cache, filter_shape, padding=None, strides=None, use_specialized_triple=True,
+                  reuse_mask=True, precomputed=None):
 
         h_filter, w_filter, d_filter, n_filter = filter_shape
         d_y_reshaped = d_y.transpose(1, 2, 3, 0).reshape(n_filter, -1)
@@ -1061,33 +1085,41 @@ class PrivateEncodedTensor:
             dw = dw.reshape(filter_shape)
             return dw
         if isinstance(d_y, PrivateEncodedTensor):
-            # we need: x, a, a_col, alpha_col
-            if reuse_mask:
-                a, a_col, alpha_col = cache
-                a, b, a_convbw_b = generate_convbw_triple(a.shape, d_y_reshaped.shape, a=a, a_col=a_col)
-                beta = (d_y_reshaped - b).reveal()
+            if use_specialized_triple:
+                # we need: x, a, a_col, alpha_col
+                if reuse_mask:
+                    a, a_col, alpha_col = x.mask, x.mask_transformed, x.masked_transformed
+                    a, b, a_convbw_b = generate_convbw_triple(a.shape, d_y_reshaped.shape, a=a, a_col=a_col)
+                    beta = (d_y_reshaped - b).reveal()
 
-                alpha_convbw_beta = beta.dot(alpha_col.transpose())
-                alpha_convbw_b = b.dot(alpha_col.transpose())
-                a_convbw_beta = beta.dot(a_col.transpose())
+                    alpha_convbw_beta = beta.dot(alpha_col.transpose())
+                    alpha_convbw_b = b.dot(alpha_col.transpose())
+                    a_convbw_beta = beta.dot(a_col.transpose())
 
-                z = (alpha_convbw_beta + alpha_convbw_b + a_convbw_beta + a_convbw_b).reshape(filter_shape)
-                return PrivateEncodedTensor.from_shares(z.shares0, z.shares1).truncate()
+                    z = (alpha_convbw_beta + alpha_convbw_b + a_convbw_beta + a_convbw_b).reshape(filter_shape)
+                    return PrivateEncodedTensor.from_shares(z.shares0, z.shares1).truncate()
 
+                else:
+                    a, b, a_convbw_b = generate_convbw_triple(x.shape, d_y_reshaped.shape)
+                    alpha = (x - a).reveal()
+                    beta = (d_y_reshaped - b).reveal()
+
+                    alpha_col = alpha.im2col(h_filter, w_filter, padding, strides)
+                    a_col = a.im2col(h_filter, w_filter, padding, strides)
+
+                    alpha_convbw_beta = beta.dot(alpha_col.transpose())
+                    alpha_convbw_b = b.dot(alpha_col.transpose())
+                    a_convbw_beta = beta.dot(a_col.transpose())
+
+                    z = (alpha_convbw_beta + alpha_convbw_b + a_convbw_beta + a_convbw_b).reshape(filter_shape)
+                    return PrivateEncodedTensor.from_shares(z.shares0, z.shares1).truncate()
             else:
-                assert False
-                # a, b, a_convbw_b = generate_convbw_triple(x.shape, filter_shape, a=mask_x)
-                # alpha = (x - a).reveal()        # (PrivateEncodedTensor - PrivateFieldTensor).reveal() = PublicFieldTensor
-                # alpha_col = alpha.im2col(h_filter, w_filter, padding, strides) # PublicFieldTensor
+                X_col = cache
+                dw = d_y_reshaped.dot(X_col.transpose())
+                dw = dw.reshape(filter_shape)
+                return dw
 
 
-
-
-
-
-
-
-            pass
 
 
 ANALYTIC_STORE = []
